Add docker publish
