+++ client/package.json
@@ -1,7 +1,7 @@
 {
   "name": "client",
   "version": "0.1.0",
-  "proxy": "http://127.0.0.1:5000/",
+  "proxy": "http://localhost:5000/",
   "private": true,
   "dependencies": {
     "@emotion/react": "^11.10.4",
@@ -22,7 +22,7 @@
   "scripts": {
     "start": "concurrently \"react-scripts start\" \"npm run server --prefix ../server\"",
     "build": "react-scripts build",
-    "test": "react-scripts test",
+    "test": "mocha",
     "eject": "react-scripts eject",
     "client": "react-scripts start",
     "server": "node ../server/bin/www"
@@ -47,6 +47,7 @@
   },
   "devDependencies": {
     "concurrently": "^7.4.0",
-    "flow-babel-webpack-plugin": "^1.1.1"
+    "flow-babel-webpack-plugin": "^1.1.1",
+    "mocha": "^10.1.0"
   }
 }
